Edit users settings
Edit users settings

+++ modules/settings/users.php
@@ -142,10 +142,66 @@ $del = isset($_POST['del']) ? $_POST['del'] : '';
 	header("location: " . $_SERVER['REQUEST_URI']);
 	exit();
 	 }
-	 ?>
+	 
+	//change login
+	$login_id = isset($_POST['login_id']) ? $_POST['login_id'] : '';
+	$login_new = isset($_POST['login_new']) ? $_POST['login_new'] : '';
+	$new_login = isset($_POST['new_login']) ? $_POST['new_login'] : '';
+	if (!empty($login_id) && $new_login == 'new_login'){
+	$db = new PDO('sqlite:dbf/nettemp.db');
+	$db->exec("UPDATE users SET login='$login_new' WHERE id='$login_id'") or die ($db->lastErrorMsg());
+	header("location: " . $_SERVER['REQUEST_URI']);
+	exit();
+	 }
+	//change mail
+	$mail_id = isset($_POST['mail_id']) ? $_POST['mail_id'] : '';
+	$mail_new = isset($_POST['mail_new']) ? $_POST['mail_new'] : '';
+	$new_mail = isset($_POST['new_mail']) ? $_POST['new_mail'] : '';
+	if (!empty($mail_id) && $new_mail == 'new_mail'){
+	$db = new PDO('sqlite:dbf/nettemp.db');
+	$db->exec("UPDATE users SET mail='$mail_new' WHERE id='$mail_id'") or die ($db->lastErrorMsg());
+	header("location: " . $_SERVER['REQUEST_URI']);
+	exit();
+	 }
+	 
+	 //change tel
+	$tel_id = isset($_POST['tel_id']) ? $_POST['tel_id'] : '';
+	$tel_new = isset($_POST['tel_new']) ? $_POST['tel_new'] : '';
+	$new_tel = isset($_POST['new_tel']) ? $_POST['new_tel'] : '';
+	if (!empty($tel_id) && $new_tel == 'new_tel'){
+	$db = new PDO('sqlite:dbf/nettemp.db');
+	$db->exec("UPDATE users SET tel='$tel_new' WHERE id='$tel_id'") or die ($db->lastErrorMsg());
+	header("location: " . $_SERVER['REQUEST_URI']);
+	exit();
+	 }
+	 
+	  //change smspin
+	$pin_id = isset($_POST['pin_id']) ? $_POST['pin_id'] : '';
+	$pin_new = isset($_POST['pin_new']) ? $_POST['pin_new'] : '';
+	$new_pin = isset($_POST['new_pin']) ? $_POST['new_pin'] : '';
+	if (!empty($pin_id) && $new_pin == 'new_pin'){
+	$db = new PDO('sqlite:dbf/nettemp.db');
+	$db->exec("UPDATE users SET smspin='$pin_new' WHERE id='$pin_id'") or die ($db->lastErrorMsg());
+	header("location: " . $_SERVER['REQUEST_URI']);
+	exit();
+	 }
+	 
+	 //change password
+	$pass1=sha1(isset($_POST['pass1']) ? $_POST['pass1'] : '');
+	$pass2=sha1(isset($_POST['pass2']) ? $_POST['pass2'] : '');
+	$pass_change = isset($_POST['pass_change']) ? $_POST['pass_change'] : '';
+	$pass_id = isset($_POST['pass_id']) ? $_POST['pass_id'] : '';
 	
-
-<?php // SQLite - usuwanie notification
+	if (!empty($pass_id) && $pass_change == "pass_change") { 
+		if ((!empty($pass1)) && (!empty($pass2)) && ($pass1 == $pass2)) {
+		$db = new PDO('sqlite:dbf/nettemp.db');
+		$db->exec("UPDATE users SET password='$pass1' WHERE id='$pass_id' ") or die ($db->lastErrorMsg());
+		$flag = "OK";
+		}	else {$flag = "ERR";
+			}
+	}
+	
+	  // SQLite - usuwanie notification
 	if (!empty($del) && ($_POST['del1'] == "del2") ){
 	$db = new PDO('sqlite:dbf/nettemp.db');
 	$db->exec("DELETE FROM users WHERE id='$del'") or die ($db->lastErrorMsg());
@@ -171,9 +227,9 @@ $del = isset($_POST['del']) ? $_POST['del'] : '';
 
     <tr>	
 	<form action="" method="post">
-	<td><input type="text" name="login" value="" class="form-control" required=""/></td>
+	<td><input type="text" name="login" size="10" maxlength="30" value="" class="form-control" required=""/></td>
 	<td><input type="password" name="pass" value="" class="form-control" required=""/></td>
-	<td><input type="email" name="mail" value="" class="form-control" required=""/></td>
+	<td><input type="email" name="mail" size="25" maxlength="50" value="" class="form-control" required=""/></td>
 	<td><input type="text" name="tel" value="" class="form-control" required=""/></td>
 	<td><input type="text" name="smspin" value="" class="form-control" required=""/></td>
 	<input type="hidden" name="add1" value="add2" />
@@ -189,14 +245,68 @@ $result = $sth->fetchAll();
 foreach ($result as $a) { 
 ?>
 	<tr>
-	<td><?php echo $a["login"];?></td>
-	<td></td>
-	<td><?php echo $a["mail"];?></td>
-	<td><?php echo $a["tel"]; ?></td>
-	<td><?php echo $a["smspin"]; ?></td>
-	<?php if ($a['login'] != 'admin') { ?>
 	<td>
-    	<form action="" method="post"> 	
+	
+	<?php if ($a["login"] != 'admin') { ?>
+	 <form action="" method="post" style="display:inline!important;">
+		<input type="text" name="login_new" size="10" maxlength="30" value="<?php echo $a["login"]; ?>" />
+		<input type="hidden" name="login_id" value="<?php echo $a["id"]; ?>" />
+		<input type="hidden" name="new_login" value="new_login"/>
+		<button class="btn btn-xs btn-success"><span class="glyphicon glyphicon-pencil"></span> </button>
+    </form>
+	<?php } else { echo $a["login"];}?>
+	</td>
+	
+	<td>
+	<form action="" method="post" style="display:inline!important;">
+	<label>New :&nbsp</label><input type="password" name="pass1" size="15" maxlength="30" value=""  required=""/>
+	<label>Repeat :&nbsp</label><input type="password" name="pass2" size="15" maxlength="30" value="" required=""/>
+	<input type="hidden" name="pass_id" value="<?php echo $a["id"]; ?>" />
+	<input type="hidden" name="pass_change" value="pass_change"/>
+	<button class="btn btn-xs btn-success"><span class="glyphicon glyphicon-pencil"></span> </button>
+	</form>
+	<?php if ($a["id"] == $pass_id && $flag == 'OK') { ?>
+		
+		<span class="label label-success">Chenged</span>
+		
+		<?php } elseif ($a["id"] == $pass_id && $flag == 'ERR') { ?>
+		
+			<span class="label label-danger">Error</span>
+		<?php } ?>
+		
+	</td>
+	
+	<td>
+	<form action="" method="post" style="display:inline!important;">
+		<input type="text" name="mail_new" size="25" maxlength="50" value="<?php echo $a["mail"]; ?>" />
+		<input type="hidden" name="mail_id" value="<?php echo $a["id"]; ?>" />
+		<input type="hidden" name="new_mail" value="new_mail"/>
+		<button class="btn btn-xs btn-success"><span class="glyphicon glyphicon-pencil"></span> </button>
+    </form>
+	</td>
+	
+	<td>
+	<form action="" method="post" style="display:inline!important;">
+		<input type="text" name="tel_new" size="15" maxlength="50" value="<?php echo $a["tel"]; ?>" />
+		<input type="hidden" name="tel_id" value="<?php echo $a["id"]; ?>" />
+		<input type="hidden" name="new_tel" value="new_tel"/>
+		<button class="btn btn-xs btn-success"><span class="glyphicon glyphicon-pencil"></span> </button>
+    </form>
+	</td>
+	
+	<td>
+	<form action="" method="post" style="display:inline!important;">
+		<input type="text" name="pin_new" size="10" maxlength="50" value="<?php echo $a["smspin"]; ?>" />
+		<input type="hidden" name="pin_id" value="<?php echo $a["id"]; ?>" />
+		<input type="hidden" name="new_pin" value="new_pin"/>
+		<button class="btn btn-xs btn-success"><span class="glyphicon glyphicon-pencil"></span> </button>
+    </form>
+
+	</td>
+	
+	<td>
+	<?php if ($a['login'] != 'admin') { ?>
+    <form action="" method="post"> 	
 	    <input type="hidden" name="del" value="<?php echo $a["id"]; ?>" />
 	    <input type="hidden" type="submit" name="del1" value="del2" />
 	    <button class="btn btn-xs btn-danger"><span class="glyphicon glyphicon-trash"></span> </button>
